implemented abstract value annotation
