+++ Annotation/Inherit.php
@@ -17,30 +17,6 @@ namespace ONGR\ElasticsearchBundle\Annotation;
  * @Annotation
  * @Target("CLASS")
  */
-final class Inherit
+final class Inherit extends AbstractValue
 {
-    /**
-     * @var array
-     */
-    public $value;
-
-    /**
-     * Constructor.
-     *
-     * @param array $values
-     *
-     * @throws \InvalidArgumentException
-     */
-    public function __construct(array $values)
-    {
-        if (is_string($values['value'])) {
-            $this->value = [$values['value']];
-        } elseif (is_array($values['value'])) {
-            $this->value = $values['value'];
-        } else {
-            throw new \InvalidArgumentException(
-                'Annotation Inherit unexpected type given. Expected string or array, given ' . gettype($values['value'])
-            );
-        }
-    }
 }
